luke recorder

+++ dm/r.sh
@@ -1,14 +1,36 @@
 #!/bin/bash
 
-if pgrep ffmpeg; then
-    if timeout 10 inotifywait --event modify ~/paperbenni/recordings; then
-        pkill ffmpeg
-        echo "|" > ~/.status
-        exit 0
-    fi
+screencast() {
+    ffmpeg -y \
+        -f x11grab \
+        -framerate 60 \
+        -s "$W"x"$H" \
+        -i :0.0+$X,$Y \
+        -f alsa -i default \
+        -r 30 \
+        -c:v libx264rgb -crf 0 -preset ultrafast -c:a flac \
+        "$HOME/paperbenni/recordings/screencast-$(date '+%y%m%d-%H%M-%S').mkv" &
+    echo $! >/tmp/recordingpid
+}
+
+killrecording() {
+    recpid="$(cat /tmp/recordingpid)"
+    # kill with SIGTERM, allowing finishing touches.
+    kill -15 "$recpid"
+    rm -f /tmp/recordingpid
+    # even after SIGTERM, ffmpeg may still run, so SIGKILL it.
+    sleep 3
+    kill -9 "$recpid"
+    exit
+}
+
+if [ -e /tmp/recordingpid ]; then
+    killrecording
+    exit
+    exit
 fi
 
 slop=$(slop -f "%x %y %w %h %g %i") || exit 1
 read -r X Y W H G ID < <(echo $slop)
-echo "r" > ~/.status
-ffmpeg -framerate 30 -f x11grab -s "$W"x"$H" -i :0.0+$X,$Y -f alsa -i pulse -vcodec libx264 ~/paperbenni/recordings/$(date '+%Y%m%d%H%M%S').mkv
+echo "r" >~/.status
+screencast
